+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- ***Breaking*** The entire api for Headless contexts have been removed. Please instead use `Context::new()` when trying to make a context without a visible window. Also removed `headless` feature.
+- ***Breaking*** Structs implementing the `GlContext` trait must now be sized.
+- ***Breaking*** Added new `CreationErrorPair` enum variant to enum `CreationError`.
+
 # Version 0.18.0 (2018-08-03)
 
 - cocoa and core-graphics updates.